+++ test/integration/failover_eventual_test.lua
@@ -357,7 +357,7 @@ g.test_api_failover = function()
 
     local tarantool_defaults = {
         uri = 'tcp://localhost:4401',
-        password = '',
+        password = '******',
     }
 
     local etcd2_params = {
@@ -365,7 +365,7 @@ g.test_api_failover = function()
         lock_delay = 10,
         endpoints = {'goo.gl:9'},
         username = '',
-        password = '',
+        password = '******',
     }
     t.assert_equals(
         set_failover_params({
@@ -378,7 +378,7 @@ g.test_api_failover = function()
                 prefix = 'kv',
                 lock_delay = 36.6,
                 username = '',
-                password = '',
+                password = '******',
                 endpoints = {
                     'http://127.0.0.1:4001',
                     'http://127.0.0.1:2379',
@@ -401,7 +401,7 @@ g.test_api_failover = function()
                 lock_delay = 10,
                 endpoints = {'goo.gl:9'},
                 username = '',
-                password = '',
+                password = '******',
             },
             fencing_enabled = false,
             fencing_timeout = 4,
@@ -423,13 +423,14 @@ g.test_api_failover = function()
     )
 
     local tarantool_params = {uri = 'stateboard.com:8', password = 'xxx'}
+    local tarantool_params_masked = {uri = 'stateboard.com:8', password = '******'}
     t.assert_equals(
         set_failover_params({tarantool_params = tarantool_params}),
         {
             mode = 'eventual',
             failover_timeout = 0,
             etcd2_params = etcd2_params,
-            tarantool_params = tarantool_params,
+            tarantool_params = tarantool_params_masked,
             fencing_enabled = false,
             fencing_timeout = 4,
             fencing_pause = 2,
@@ -442,7 +443,7 @@ g.test_api_failover = function()
             state_provider = 'tarantool',
             failover_timeout = 0,
             etcd2_params = etcd2_params,
-            tarantool_params = tarantool_params,
+            tarantool_params = tarantool_params_masked,
             fencing_enabled = false,
             fencing_timeout = 4,
             fencing_pause = 2,
@@ -456,7 +457,7 @@ g.test_api_failover = function()
             state_provider = 'tarantool',
             failover_timeout = 0,
             etcd2_params = etcd2_params,
-            tarantool_params = tarantool_params,
+            tarantool_params = tarantool_params_masked,
             fencing_enabled = false,
             fencing_timeout = 4,
             fencing_pause = 2,
@@ -469,7 +470,7 @@ g.test_api_failover = function()
             state_provider = 'tarantool',
             failover_timeout = 0,
             etcd2_params = etcd2_params,
-            tarantool_params = tarantool_params,
+            tarantool_params = tarantool_params_masked,
             fencing_enabled = false,
             fencing_timeout = 4,
             fencing_pause = 2,
@@ -488,7 +489,7 @@ g.test_api_failover = function()
         prefix = '/',
         lock_delay = 10,
         username = '',
-        password = '',
+        password = '******',
         endpoints = {
             'http://127.0.0.1:4001',
             'http://127.0.0.1:2379',
@@ -502,7 +503,7 @@ g.test_api_failover = function()
             state_provider = 'tarantool',
             failover_timeout = 3,
             etcd2_params = etcd2_defaults,
-            tarantool_params = tarantool_params,
+            tarantool_params = tarantool_params_masked,
             fencing_enabled = false,
             fencing_timeout = 4,
             fencing_pause = 2,
@@ -515,7 +516,7 @@ g.test_api_failover = function()
             state_provider = 'tarantool',
             failover_timeout = 3,
             etcd2_params = etcd2_defaults,
-            tarantool_params = tarantool_params,
+            tarantool_params = tarantool_params_masked,
             fencing_enabled = false,
             fencing_timeout = 4,
             fencing_pause = 2,