+++ CHANGELOG.rst
@@ -26,6 +26,8 @@ Changed
 
 - Disabled instances won't appear as leaders.
 
+- Mask failover password in WebUI and GraphQL API.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Added
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~